feat: finished contextualizer
